refactoring runmat-cli for maintainability
