+++ crates/runmat-cli/src/app/mod.rs
@@ -1,0 +1,2 @@
+pub mod bootstrap;
+pub mod dispatch;
