+++ crates/runmat-cli/src/commands/mod.rs
@@ -1,0 +1,14 @@
+pub mod accel;
+pub mod benchmark;
+pub mod bytecode;
+pub mod config;
+pub mod gc;
+pub mod jupyter;
+pub mod kernel;
+pub mod pkg;
+pub mod repl;
+pub mod script;
+pub mod session;
+pub mod snapshot;
+pub mod streams;
+pub mod version;
