+++ crates/runmat-config/src/lib.rs
@@ -184,9 +184,26 @@ pub struct TelemetryConfig {
     /// Bounded queue size for async delivery
     #[serde(default = "default_telemetry_queue")]
     pub queue_size: usize,
+    /// Deliver telemetry synchronously on the caller thread
+    #[serde(default)]
+    pub sync_mode: bool,
+    /// Drain policy used when the process exits
+    #[serde(default)]
+    pub drain_mode: TelemetryDrainMode,
+    /// Maximum time to wait for telemetry drain on shutdown
+    #[serde(default = "default_telemetry_drain_timeout_ms")]
+    pub drain_timeout_ms: u64,
     /// Require ingestion key (self-built binaries default to false)
     #[serde(default = "default_true")]
     pub require_ingestion_key: bool,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, ValueEnum, Default)]
+#[serde(rename_all = "kebab-case")]
+pub enum TelemetryDrainMode {
+    None,
+    #[default]
+    All,
 }
 
 impl Default for TelemetryConfig {
@@ -197,6 +214,9 @@ impl Default for TelemetryConfig {
             http_endpoint: None,
             udp_endpoint: Some("udp.telemetry.runmat.com:7846".to_string()),
             queue_size: default_telemetry_queue(),
+            sync_mode: false,
+            drain_mode: TelemetryDrainMode::All,
+            drain_timeout_ms: default_telemetry_drain_timeout_ms(),
             require_ingestion_key: true,
         }
     }
@@ -217,6 +237,10 @@ impl Default for AccelerateConfig {
 
 fn default_telemetry_queue() -> usize {
     256
+}
+
+fn default_telemetry_drain_timeout_ms() -> u64 {
+    50
 }
 
 /// Auto-offload planner configuration
@@ -1016,6 +1040,19 @@ impl ConfigLoader {
                 config.telemetry.queue_size = parsed.max(MIN_QUEUE_SIZE);
             }
         }
+        if let Some(sync_mode) = env_bool("RUNMAT_TELEMETRY_SYNC", &[]) {
+            config.telemetry.sync_mode = sync_mode;
+        }
+        if let Some(drain_mode) = env_value("RUNMAT_TELEMETRY_DRAIN", &[]) {
+            if let Some(parsed) = parse_telemetry_drain_mode(&drain_mode) {
+                config.telemetry.drain_mode = parsed;
+            }
+        }
+        if let Some(drain_timeout) = env_value("RUNMAT_TELEMETRY_DRAIN_TIMEOUT_MS", &[]) {
+            if let Ok(parsed) = drain_timeout.trim().parse::<u64>() {
+                config.telemetry.drain_timeout_ms = parsed;
+            }
+        }
 
         // Acceleration settings
         if let Some(accel) = env_value("RUNMAT_ACCEL_ENABLE", &[]) {
@@ -1251,6 +1288,16 @@ fn parse_power_preference(value: &str) -> Option<AccelPowerPreference> {
     }
 }
 
+fn parse_telemetry_drain_mode(value: &str) -> Option<TelemetryDrainMode> {
+    match value.trim().to_ascii_lowercase().as_str() {
+        "started" | "start" | "session" | "all" | "full" | "both" | "runtime" => {
+            Some(TelemetryDrainMode::All)
+        }
+        "none" | "off" | "" => Some(TelemetryDrainMode::None),
+        _ => None,
+    }
+}
+
 fn env_value(primary: &str, aliases: &[&str]) -> Option<String> {
     env::var(primary)
         .ok()
@@ -1427,5 +1474,24 @@ mod tests {
         assert!(config.telemetry.udp_endpoint.is_none());
         std::env::remove_var("RUNMAT_TELEMETRY_ENDPOINT");
         std::env::remove_var("RUNMAT_TELEMETRY_UDP_ENDPOINT");
+    }
+
+    #[test]
+    fn telemetry_runtime_env_overrides_promote_into_config() {
+        let _lock = ENV_GUARD.lock().unwrap();
+        std::env::set_var("RUNMAT_TELEMETRY_SYNC", "1");
+        std::env::set_var("RUNMAT_TELEMETRY_DRAIN", "none");
+        std::env::set_var("RUNMAT_TELEMETRY_DRAIN_TIMEOUT_MS", "250");
+
+        let mut config = RunMatConfig::default();
+        ConfigLoader::apply_environment_variables(&mut config).unwrap();
+
+        assert!(config.telemetry.sync_mode);
+        assert_eq!(config.telemetry.drain_mode, TelemetryDrainMode::None);
+        assert_eq!(config.telemetry.drain_timeout_ms, 250);
+
+        std::env::remove_var("RUNMAT_TELEMETRY_SYNC");
+        std::env::remove_var("RUNMAT_TELEMETRY_DRAIN");
+        std::env::remove_var("RUNMAT_TELEMETRY_DRAIN_TIMEOUT_MS");
     }
 }
